requirements.txt
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.3
numpy==2.3.2
pandas==2.3.2
scikit-learn==1.7.1
requests==2.32.5
SQLAlchemy==2.0.43
psutil
python-dotenv==1.1.1

